poa tx payload builder fixes & epoch manager fixes

+++ crates/consensus/common/src/utils.rs
@@ -216,8 +216,8 @@ pub fn validate_against_parent(
     let current_signer = recovery_authority(&current)
         .map_err(|e| ValidateAgainstParentError::FailedToDerserializeExtraData(e))?;
     // Check if the parent block was mined in a different turn
-    let parent_ts = parent.timestamp / 60;
-    let current_ts = current.timestamp / 60;
+    let parent_ts = parent.timestamp as f64 / 60.0;
+    let current_ts = current.timestamp as f64 / 60.0;
 
     validate_current_signer_against_last((parent_signer, parent_ts), (current_signer, current_ts))?;
 
@@ -227,13 +227,13 @@ pub fn validate_against_parent(
 /// Validate current signer and its last block timestamp against the last signer and its last block
 /// timestamp Used to prevent a signer from signing multiple blocks in the same turn
 pub fn validate_current_signer_against_last(
-    last: (secp256k1::PublicKey, u64),
-    current: (secp256k1::PublicKey, u64),
+    last: (secp256k1::PublicKey, f64),
+    current: (secp256k1::PublicKey, f64),
 ) -> Result<(), ValidateAgainstParentError> {
     // Last block should be greater that 1 minute in the worst cast
     // Even in the case of > 2 federation members the worst case time between blocks for the same
     // Signer should be 1 minute. Assuming 1 minute block times
-    if last.0 == current.0 && current.1 - last.1 < 1 {
+    if last.0 == current.0 && current.1 - last.1 < 1.0 {
         return Err(ValidateAgainstParentError::SignerLimitExceeded);
     }
